obs-lab/agent-service/requirements.txt
flask
kafka-python
requests